+++ UnitTests/Core/InteropServices/SafeHGlobalHandleTests.cs
@@ -15,24 +15,24 @@ namespace Vanara.InteropServices.Tests
 			var r = new[] { 5, 5, 5, 5 };
 			var h = SafeHGlobalHandle.CreateFromList(r, r.Length);
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(Marshal.SizeOf(typeof(int)) * r.Length));
+			Assert.That((int)h.Size, Is.EqualTo(Marshal.SizeOf(typeof(int)) * r.Length));
 			Assert.That(h.ToStructure<int>(), Is.EqualTo(5));
 			Assert.That(h.ToEnumerable<int>(4), Has.Exactly(4).EqualTo(5).And.Exactly(4).Items);
 
-			var d = new[] { new RECT(1,1,1,1), new RECT(2,2,2,2) };
+			var d = new[] { new RECT(1, 1, 1, 1), new RECT(2, 2, 2, 2) };
 			h = SafeHGlobalHandle.CreateFromList(d, d.Length);
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(Marshal.SizeOf(typeof(RECT)) * d.Length));
+			Assert.That((int)h.Size, Is.EqualTo(Marshal.SizeOf(typeof(RECT)) * d.Length));
 			Assert.That(h.ToStructure<RECT>().X, Is.EqualTo(1));
 			Assert.That(h.ToArray<RECT>(4), Has.Exactly(4).Items);
 
-			var p = new[] { new PRECT(1,1,1,1), new PRECT(2,2,2,2) };
+			var p = new[] { new PRECT(1, 1, 1, 1), new PRECT(2, 2, 2, 2) };
 			h = SafeHGlobalHandle.CreateFromList(p);
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(Marshal.SizeOf(typeof(PRECT)) * p.Length));
+			Assert.That((int)h.Size, Is.EqualTo(Marshal.SizeOf(typeof(PRECT)) * p.Length));
 			Assert.That(h.ToArray<RECT>(4), Has.Exactly(4).Items);
 
-			Assert.That(() => SafeHGlobalHandle.CreateFromList(new[] {"X"}), Throws.ArgumentException);
+			Assert.That(() => SafeHGlobalHandle.CreateFromList(new[] { "X" }), Throws.ArgumentException);
 		}
 
 		[Test(Description = "Allocate an enumeration of strings.")]
@@ -42,33 +42,33 @@ namespace Vanara.InteropServices.Tests
 
 			var h = SafeHGlobalHandle.CreateFromStringList(r, StringListPackMethod.Concatenated, CharSet.Ansi, 7);
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(16));
+			Assert.That((int)h.Size, Is.EqualTo(16));
 			Assert.That(h.ToStringEnum(CharSet.Ansi, 7), Has.Exactly(4).EqualTo("5").And.Exactly(4).Items);
 			h = SafeHGlobalHandle.CreateFromStringList(r, StringListPackMethod.Concatenated, CharSet.Unicode, 7);
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(25));
+			Assert.That((int)h.Size, Is.EqualTo(25));
 			Assert.That(h.ToStringEnum(CharSet.Unicode, 7), Has.Exactly(4).EqualTo("5").And.Exactly(4).Items);
 			h = SafeHGlobalHandle.CreateFromStringList(r);
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(18));
+			Assert.That((int)h.Size, Is.EqualTo(18));
 			Assert.That(h.ToString(-1), Is.EqualTo("5"));
 			Assert.That(h.ToStringEnum(), Has.Exactly(4).EqualTo("5").And.Exactly(4).Items);
 
 			h = SafeHGlobalHandle.CreateFromStringList(r, StringListPackMethod.Packed, CharSet.Ansi, 7);
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(7 + IntPtr.Size + r.Length * (2 + IntPtr.Size)));
+			Assert.That((int)h.Size, Is.EqualTo(7 + IntPtr.Size + r.Length * (2 + IntPtr.Size)));
 			Assert.That(h.ToStringEnum(4, CharSet.Ansi, 7), Has.Exactly(4).EqualTo("5").And.Exactly(4).Items);
 			h = SafeHGlobalHandle.CreateFromStringList(r, StringListPackMethod.Packed, CharSet.Unicode, 7);
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(7 + IntPtr.Size + r.Length * (4 + IntPtr.Size)));
+			Assert.That((int)h.Size, Is.EqualTo(7 + IntPtr.Size + r.Length * (4 + IntPtr.Size)));
 			Assert.That(h.ToStringEnum(4, CharSet.Unicode, 7), Has.Exactly(4).EqualTo("5").And.Exactly(4).Items);
 			h = SafeHGlobalHandle.CreateFromStringList(r, StringListPackMethod.Packed);
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(IntPtr.Size + r.Length * (4 + IntPtr.Size)));
+			Assert.That((int)h.Size, Is.EqualTo(IntPtr.Size + r.Length * (4 + IntPtr.Size)));
 			Assert.That(h.ToStringEnum(4, CharSet.Unicode), Has.Exactly(4).EqualTo("5").And.Exactly(4).Items);
 
 			h = SafeHGlobalHandle.CreateFromStringList(null);
-			Assert.That(h.Size, Is.EqualTo(Extensions.StringHelper.GetCharSize()));
+			Assert.That((int)h.Size, Is.EqualTo(Extensions.StringHelper.GetCharSize()));
 		}
 
 		[Test(Description = "Allocate a structure")]
@@ -77,7 +77,7 @@ namespace Vanara.InteropServices.Tests
 			var r = new RECT(5, 5, 5, 5);
 			var h = SafeHGlobalHandle.CreateFromStructure(r);
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(Marshal.SizeOf(typeof(RECT))));
+			Assert.That((int)h.Size, Is.EqualTo(Marshal.SizeOf(typeof(RECT))));
 			Assert.That(h.ToStructure<RECT>().X, Is.EqualTo(5));
 		}
 
@@ -97,7 +97,7 @@ namespace Vanara.InteropServices.Tests
 			var ptr = Marshal.AllocHGlobal(5);
 			var h = new SafeHGlobalHandle(ptr, 5);
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(5));
+			Assert.That((int)h.Size, Is.EqualTo(5));
 			h.Dispose();
 			Assert.That(h.IsClosed && h.IsInvalid);
 
@@ -125,7 +125,7 @@ namespace Vanara.InteropServices.Tests
 
 			Assert.That(() => new SafeHGlobalHandle(-1), Throws.TypeOf<ArgumentOutOfRangeException>());
 
-			Assert.That(() => new SafeHGlobalHandle(5).AddSubReference(new [] { new SafeHGlobalHandle(5), new SafeHGlobalHandle(5) }), Throws.Nothing);
+			Assert.That(() => new SafeHGlobalHandle(5).AddSubReference(new[] { new SafeHGlobalHandle(5), new SafeHGlobalHandle(5) }), Throws.Nothing);
 		}
 
 		[Test()]
@@ -133,7 +133,7 @@ namespace Vanara.InteropServices.Tests
 		{
 			var h = new SafeHGlobalHandle(Marshal.AllocHGlobal(5), 5);
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(5));
+			Assert.That((int)h.Size, Is.EqualTo(5));
 			h.Dispose();
 			Assert.That(h.IsClosed && h.IsInvalid);
 		}
@@ -143,7 +143,7 @@ namespace Vanara.InteropServices.Tests
 		{
 			var h = new SafeHGlobalHandle(new byte[] { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 });
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(10));
+			Assert.That((int)h.Size, Is.EqualTo(10));
 			h.Dispose();
 			Assert.That(h.IsClosed && h.IsInvalid);
 		}
@@ -153,7 +153,7 @@ namespace Vanara.InteropServices.Tests
 		{
 			var h = new SafeHGlobalHandle(new[] { (IntPtr)1, (IntPtr)2, (IntPtr)3, (IntPtr)4, (IntPtr)5 });
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(5 * IntPtr.Size));
+			Assert.That((int)h.Size, Is.EqualTo(5 * IntPtr.Size));
 			h.Dispose();
 			Assert.That(h.IsClosed && h.IsInvalid);
 		}
@@ -163,7 +163,7 @@ namespace Vanara.InteropServices.Tests
 		{
 			var h = new SafeHGlobalHandle("0123456789");
 			Assert.That(!h.IsClosed && !h.IsInvalid);
-			Assert.That(h.Size, Is.EqualTo(11 * Marshal.SystemDefaultCharSize));
+			Assert.That((int)h.Size, Is.EqualTo(11 * Marshal.SystemDefaultCharSize));
 			h.Dispose();
 			Assert.That(h.IsClosed && h.IsInvalid);
 		}
